requirements.txt
absl-py==0.8.0
appnope==0.1.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
dlib==19.18.0
efficientnet==0.0.4
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
imutils==0.5.3
ipykernel==5.1.2
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
multidict==4.5.2
networkx==2.3
numpy==1.17.2
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.3.0
